Update GeoIP database installation to account for change in MaxMind's policies.
MaxMind now requires a license key for both the paid GeoIP2 and the free
GeoLite2 databases. The old download URL for the GeoLite2 database no longer
works, so any GeoLite2 tasks must be removed. Now the GeoIP2 tasks will only
run if a license key is found.

The usage of strings for the MaxMind URL building has been completely revamped
to improve readability and isolate some of the variables to just the parameter.