Install cisagov/ansible-role-geoip2 in the role tasks
This is done instead of relying on importing via the metadata
dependencies.  This allows us to set a role variable for GeoIP2 using
the role variables for this role.

Note that roles listed in meta/main.yml are installed before the
current role, whereas meta/requirements.yml are merely made available
to the current role.  This is a key difference that we exploit.